fix: correcting datatypes

+++ src/util.c
@@ -826,13 +826,13 @@ int savefai(const char *uuid, int64_t exp, const char *tok)
 
 int hextobyte(const char *hexstr, unsigned char *bytes, int64_t sz)
 {
-    int len = strlen(hexstr);
+    int64_t len = strlen(hexstr);
     if (len % 2)
         return -1;
     len /= 2;
     if (len != sz)
         return -1;
-    int i = 0;
+    int64_t i = 0;
     for (; i < len; i++)
         sscanf(hexstr + i * HEXDIG_LEN, "%2hhx", bytes + i);
     return 0;
